requirements.txt
asgiref>=3.8.1
certifi>=2024.2.2
charset-normalizer>=3.3.2
Django>=5.1,<6.0
django-cors-headers>=4.3.1
django-environ>=0.11.2
django-filter>=24.2
djangorestframework>=3.15.0
et_xmlfile==2.0.0
idna==3.11
openpyxl==3.1.5
pillow==12.1.1
psycopg2-binary==2.9.11
requests==2.32.5
sqlparse==0.5.5
urllib3==2.6.3